feat(stories): remove dublicate story assets getter
