+++ src/scripts/components/stories/story-media/story-media.tsx
@@ -5,7 +5,7 @@ import {PreviousIcon} from '../../main/icons/previous-icon';
 import {NextIcon} from '../../main/icons/next-icon';
 import {FullscreenExitIcon} from '../../main/icons/fullscreen-exit-icon';
 import {FullscreenIcon} from '../../main/icons/fullscreen-icon';
-import {getStoryMediaUrl} from '../../../libs/get-story-urls';
+import {getStoryAssetUrl} from '../../../libs/get-story-asset-urls';
 
 import styles from './story-media.styl';
 
@@ -73,7 +73,7 @@ const StoryMedia: FunctionComponent<Props> = ({
           }}>
           {images.map((image, index) => {
             const imageCaption = imageCaptions?.find((_, i) => i === index);
-            const imageUrl = getStoryMediaUrl(storyId, image);
+            const imageUrl = getStoryAssetUrl(storyId, image);
 
             return (
               <div
